requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
firebase-admin==6.4.0
python-multipart==0.0.9
pydantic==2.6.1
python-jose==3.3.0
openai==1.14.0
PyPDF2==3.0.1
httpx==0.27.2 